move handle_try_wake
