+++ proxy/src/proxy/wake_compute.rs
@@ -2,10 +2,11 @@ use crate::auth::backend::{ComputeCredentialKeys, ComputeCredentials, ComputeUse
 use crate::console::errors::WakeComputeError;
 use crate::console::{self, provider::CachedNodeInfo};
 use crate::context::RequestMonitoring;
-use crate::proxy::connect_compute::handle_try_wake;
 use crate::proxy::retry::retry_after;
 use std::ops::ControlFlow;
 use tracing::{error, warn};
+
+use super::retry::ShouldRetry;
 
 /// wake a compute (or retrieve an existing compute session from cache)
 pub async fn wake_compute(
@@ -41,4 +42,24 @@ pub async fn wake_compute(
     };
 
     Ok((node, compute_credentials.info))
+}
+
+/// Attempts to wake up the compute node.
+/// * Returns Ok(Continue(e)) if there was an error waking but retries are acceptable
+/// * Returns Ok(Break(node)) if the wakeup succeeded
+/// * Returns Err(e) if there was an error
+pub fn handle_try_wake(
+    result: Result<console::CachedNodeInfo, WakeComputeError>,
+    num_retries: u32,
+) -> Result<ControlFlow<console::CachedNodeInfo, WakeComputeError>, WakeComputeError> {
+    match result {
+        Err(err) => match &err {
+            WakeComputeError::ApiError(api) if api.should_retry(num_retries) => {
+                Ok(ControlFlow::Continue(err))
+            }
+            _ => Err(err),
+        },
+        // Ready to try again.
+        Ok(new) => Ok(ControlFlow::Break(new)),
+    }
 }
