+++ proxy/src/serverless.rs
@@ -3,9 +3,9 @@
 //! Handles both SQL over HTTP and SQL over Websockets.
 
 mod conn_pool;
+mod json;
 mod sql_over_http;
 mod websocket;
-mod json;
 
 pub use conn_pool::GlobalConnPoolOptions;
